Corrections in lines 5-8 and 28.
When I first uploaded this in the wee hours of the night, I didn't catch an issue in the first few lines. After the user inputs multiple maxes for x, x would store the last input entered. What should have happened is it should store them separately, like a list, than one variable. Now, this issue is fixed and runs as it should.